Support `asakusafw.sdk.*` convention on Gradle.
The following convention property is also now available:
* `asakusafw.vanilla.version`
  * Asakusa Vanilla libraries version.
